% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.R
\name{withAnim}
\alias{withAnim}
\title{This function needs to be added in the UI if to want to add animation to your UI elements using shinyanimation.}
\usage{
withAnim()
}
\description{
This function needs to be added in the UI if to want to add animation to your UI elements using shinyanimation.
}
\examples{
\dontrun{withAnim()}
}
\seealso{
\code{\link{startAnim}}
}
