% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anim.R
\name{withAnim}
\alias{withAnim}
\title{Set up Shiny app to use animation}
\usage{
withAnim()
}
\description{
This function needs to be added in the UI if to want to add animation to your UI elements using shinyanimate.
}
\examples{
if(interactive()){withAnim()}
}
\seealso{
\code{\link{startAnim}}
}
